+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [#859](https://github.com/estruyf/vscode-front-matter/issues/859): Fix label in the data view dropdown field
 - [#876](https://github.com/estruyf/vscode-front-matter/issues/876): Fix snippet type on the snippet card
 - [#879](https://github.com/estruyf/vscode-front-matter/issues/879): Fix for auto updating last modified date on save
+- [#885](https://github.com/estruyf/vscode-front-matter/issues/885): Fix content relationship for none i18n content
 
 ## [10.5.1] - 2024-10-23